remove `extensions/storage` leftovers
