+++ extensions/appprovider/pkg/command/command.go
@@ -12,7 +12,7 @@ import (
 	"github.com/oklog/run"
 	"github.com/owncloud/ocis/extensions/appprovider/pkg/config"
 	"github.com/owncloud/ocis/extensions/appprovider/pkg/config/parser"
-	"github.com/owncloud/ocis/extensions/storage/pkg/server/debug"
+	"github.com/owncloud/ocis/extensions/appprovider/pkg/server/debug"
 	ociscfg "github.com/owncloud/ocis/ocis-pkg/config"
 	"github.com/owncloud/ocis/ocis-pkg/log"
 	"github.com/owncloud/ocis/ocis-pkg/sync"
@@ -63,13 +63,9 @@ func AppProvider(cfg *config.Config) *cli.Command {
 			})
 
 			debugServer, err := debug.Server(
-				debug.Name(c.Command.Name+"-debug"),
-				debug.Addr(cfg.Debug.Addr),
 				debug.Logger(logger),
 				debug.Context(ctx),
-				debug.Pprof(cfg.Debug.Pprof),
-				debug.Zpages(cfg.Debug.Zpages),
-				debug.Token(cfg.Debug.Token),
+				debug.Config(cfg),
 			)
 
 			if err != nil {
