+++ extensions/frontend/pkg/command/command.go
@@ -13,7 +13,7 @@ import (
 	"github.com/oklog/run"
 	"github.com/owncloud/ocis/extensions/frontend/pkg/config"
 	"github.com/owncloud/ocis/extensions/frontend/pkg/config/parser"
-	"github.com/owncloud/ocis/extensions/storage/pkg/server/debug"
+	"github.com/owncloud/ocis/extensions/frontend/pkg/server/debug"
 	ociscfg "github.com/owncloud/ocis/ocis-pkg/config"
 	"github.com/owncloud/ocis/ocis-pkg/log"
 	"github.com/owncloud/ocis/ocis-pkg/sync"
@@ -107,13 +107,9 @@ func Frontend(cfg *config.Config) *cli.Command {
 
 			{
 				server, err := debug.Server(
-					debug.Name(c.Command.Name+"-debug"),
-					debug.Addr(cfg.Debug.Addr),
 					debug.Logger(logger),
 					debug.Context(ctx),
-					debug.Pprof(cfg.Debug.Pprof),
-					debug.Zpages(cfg.Debug.Zpages),
-					debug.Token(cfg.Debug.Token),
+					debug.Config(cfg),
 				)
 
 				if err != nil {
