+++ extensions/ocdav/pkg/command/ocdav.go
@@ -9,7 +9,7 @@ import (
 	"github.com/oklog/run"
 	"github.com/owncloud/ocis/extensions/ocdav/pkg/config"
 	"github.com/owncloud/ocis/extensions/ocdav/pkg/config/parser"
-	"github.com/owncloud/ocis/extensions/storage/pkg/server/debug"
+	"github.com/owncloud/ocis/extensions/ocdav/pkg/server/debug"
 	ociscfg "github.com/owncloud/ocis/ocis-pkg/config"
 	"github.com/owncloud/ocis/ocis-pkg/log"
 	"github.com/owncloud/ocis/ocis-pkg/sync"
@@ -77,13 +77,9 @@ func OCDav(cfg *config.Config) *cli.Command {
 
 			{
 				server, err := debug.Server(
-					debug.Name(c.Command.Name+"-debug"),
-					debug.Addr(cfg.Debug.Addr),
 					debug.Logger(logger),
 					debug.Context(ctx),
-					debug.Pprof(cfg.Debug.Pprof),
-					debug.Zpages(cfg.Debug.Zpages),
-					debug.Token(cfg.Debug.Token),
+					debug.Config(cfg),
 				)
 
 				if err != nil {
