+++ extensions/storage-publiclink/pkg/command/storagepubliclink.go
@@ -12,7 +12,7 @@ import (
 	"github.com/oklog/run"
 	"github.com/owncloud/ocis/extensions/storage-publiclink/pkg/config"
 	"github.com/owncloud/ocis/extensions/storage-publiclink/pkg/config/parser"
-	"github.com/owncloud/ocis/extensions/storage/pkg/server/debug"
+	"github.com/owncloud/ocis/extensions/storage-publiclink/pkg/server/debug"
 	ociscfg "github.com/owncloud/ocis/ocis-pkg/config"
 	"github.com/owncloud/ocis/ocis-pkg/log"
 	"github.com/owncloud/ocis/ocis-pkg/sync"
@@ -66,13 +66,9 @@ func StoragePublicLink(cfg *config.Config) *cli.Command {
 			})
 
 			debugServer, err := debug.Server(
-				debug.Name(c.Command.Name+"-debug"),
-				debug.Addr(cfg.Debug.Addr),
 				debug.Logger(logger),
 				debug.Context(ctx),
-				debug.Pprof(cfg.Debug.Pprof),
-				debug.Zpages(cfg.Debug.Zpages),
-				debug.Token(cfg.Debug.Token),
+				debug.Config(cfg),
 			)
 
 			if err != nil {
