+++ extensions/storage-shares/pkg/command/command.go
@@ -16,7 +16,7 @@ import (
 	"github.com/oklog/run"
 	"github.com/owncloud/ocis/extensions/storage-shares/pkg/config"
 	"github.com/owncloud/ocis/extensions/storage-shares/pkg/config/parser"
-	"github.com/owncloud/ocis/extensions/storage/pkg/server/debug"
+	"github.com/owncloud/ocis/extensions/storage-shares/pkg/server/debug"
 	ociscfg "github.com/owncloud/ocis/ocis-pkg/config"
 	"github.com/thejerf/suture/v4"
 	"github.com/urfave/cli/v2"
@@ -68,13 +68,9 @@ func StorageShares(cfg *config.Config) *cli.Command {
 			})
 
 			debugServer, err := debug.Server(
-				debug.Name(c.Command.Name+"-debug"),
-				debug.Addr(cfg.Debug.Addr),
 				debug.Logger(logger),
 				debug.Context(ctx),
-				debug.Pprof(cfg.Debug.Pprof),
-				debug.Zpages(cfg.Debug.Zpages),
-				debug.Token(cfg.Debug.Token),
+				debug.Config(cfg),
 			)
 
 			if err != nil {
