--- extensions/storage/.dockerignore
@@ -1,2 +1,0 @@
-*
-!bin/
