--- extensions/storage/Makefile
@@ -1,37 +1,0 @@
-SHELL := bash
-NAME := storage
-
-include ../../.make/recursion.mk
-
-############ tooling ############
-ifneq (, $(shell which go 2> /dev/null)) # suppress `command not found warnings` for non go targets in CI
-include ../../.bingo/Variables.mk
-endif
-
-############ go tooling ############
-include ../../.make/go.mk
-
-############ release #########
-include ../../.make/release.mk
-
-############ docs generate ############
-include ../../.make/docs.mk
-
-.PHONY: docs-generate
-docs-generate: config-docs-generate
-
-############ generate ############
-include ../../.make/generate.mk
-
-.PHONY: ci-go-generate
-ci-go-generate: # CI runs ci-node-generate automatically before this target
-
-.PHONY: ci-node-generate
-ci-node-generate:
-
-############ licenses ############
-.PHONY: ci-node-check-licenses
-ci-node-check-licenses:
-
-.PHONY: ci-node-save-licenses
-ci-node-save-licenses:
