--- extensions/storage/docker/Dockerfile.linux.arm
@@ -1,19 +1,0 @@
-FROM arm32v6/alpine:latest
-
-RUN apk update && \
-	apk upgrade && \
-	apk add ca-certificates mailcap && \
-	rm -rf /var/cache/apk/* && \
-	echo 'hosts: files dns' >| /etc/nsswitch.conf
-
-LABEL maintainer="ownCloud GmbH <devops@owncloud.com>" \
-  org.label-schema.name="oCIS Reva" \
-  org.label-schema.vendor="ownCloud GmbH" \
-  org.label-schema.schema-version="1.0"
-
-EXPOSE 9140 9141 9142 9143
-
-ENTRYPOINT ["/usr/bin/ocis-reva"]
-CMD ["server"]
-
-COPY bin/ocis-reva /usr/bin/ocis-reva
