--- extensions/storage/pkg/server/debug/server.go
@@ -1,59 +1,0 @@
-package debug
-
-import (
-	"io"
-	"net/http"
-
-	"github.com/owncloud/ocis/extensions/storage/pkg/config"
-	"github.com/owncloud/ocis/ocis-pkg/service/debug"
-	"github.com/owncloud/ocis/ocis-pkg/version"
-)
-
-// Server initializes the debug service and server.
-func Server(opts ...Option) (*http.Server, error) {
-	options := newOptions(opts...)
-
-	return debug.NewService(
-		debug.Logger(options.Logger),
-		debug.Name(options.Name),
-		debug.Version(version.String),
-		debug.Address(options.Addr),
-		debug.Token(options.Token),
-		debug.Pprof(options.Pprof),
-		debug.Zpages(options.Zpages),
-		debug.Health(health(options.Config)),
-		debug.Ready(ready(options.Config)),
-	), nil
-}
-
-// health implements the health check.
-func health(cfg *config.Config) func(http.ResponseWriter, *http.Request) {
-	return func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "text/plain")
-		w.WriteHeader(http.StatusOK)
-
-		// TODO: check if services are up and running
-
-		_, err := io.WriteString(w, http.StatusText(http.StatusOK))
-		// io.WriteString should not fail but if it does we want to know.
-		if err != nil {
-			panic(err)
-		}
-	}
-}
-
-// ready implements the ready check.
-func ready(cfg *config.Config) func(http.ResponseWriter, *http.Request) {
-	return func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "text/plain")
-		w.WriteHeader(http.StatusOK)
-
-		// TODO: check if services are up and running
-
-		_, err := io.WriteString(w, http.StatusText(http.StatusOK))
-		// io.WriteString should not fail but if it does we want to know.
-		if err != nil {
-			panic(err)
-		}
-	}
-}
