--- extensions/storage/reflex.conf
@@ -1,5 +1,0 @@
-# backend
--r '^(cmd|pkg)/.*\.go$' -R '^node_modules/' -s -- sh -c 'make bin/ocis-reva-debug && bin/ocis-reva-debug --log-level debug server --debug-pprof --debug-zpages --asset-path assets/'
-
-# frontend
--r '^ui/.*\.(vue|js)$' -R '^node_modules/' -- sh -c 'yarn build'
